requirements.txt
# ============================================================
# Agentic Maintenance Planning System - Dependencies
# ============================================================

# LangGraph & LangChain
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# LangGraph Checkpointer for PostgreSQL
langgraph-checkpoint-postgres>=2.0.0

# PostgreSQL Driver (psycopg3 - required by LangGraph checkpoint)
psycopg[binary]>=3.1.0
psycopg-pool>=3.2.0

# OpenAI
openai>=1.50.0

# Chainlit UI
chainlit>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
